requirements.txt
# ─── PHASAL Vision AI – Dependencies ─────────────────────────
# Install with: pip install -r requirements.txt

# Core ML frameworks
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
safetensors>=0.3.0

# Web frameworks
flask>=3.0.0
streamlit>=1.28.0

# Image processing
Pillow>=9.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
